added check to see if entry actually exists

+++ admin_tools/remove_entries.php
@@ -14,16 +14,7 @@ $path = $_SERVER['DOCUMENT_ROOT'].$root; ?>
 <h3>Remove A Location</h3>	
 </div>
 	<?php 
-		//Check if samples still exist for location
-		//if samples exists, tell user they have to update the location name for each of the other samples
-		//(bulk update?)
-		//list out the samples for the user (xls?)
-		
-		//if there are no samples for this location name, go ahead and change visible to 0
 
-	
-	
-	
 		//error && type checking 
 		if(isset($_GET['submit']) && $_GET['submit'] == 'Delete'){
 			
@@ -45,35 +36,68 @@ $path = $_SERVER['DOCUMENT_ROOT'].$root; ?>
 			}
 			
 			if($entry_type == 'dExtrKit'){
+				$select_exists_query = "SELECT d_kit_name FROM dna_extraction WHERE d_kit_name = ?";
 				$select_query = "SELECT sample_name FROM sample WHERE dna_extract_kit_name = ?";
 				$update_query = "UPDATE dna_extraction SET visible = ? WHERE d_kit_name = ? ";
 			}
 			if($entry_type == 'Location'){
+				$select_exists_query = "SELECT loc_name FROM location WHERE loc_name = ?";
 				$select_query = "SELECT sample_name FROM sample WHERE location_name = ?";
 				$update_query = "UPDATE location SET visible = ? WHERE loc_name = ? ";
 			}
 			if($entry_type == 'Media Type'){
+				$select_exists_query = "SELECT media_type FROM media_type WHERE media_type = ?";
 				$select_query = "SELECT sample_name FROM sample WHERE media_type = ?";
 				$update_query = "UPDATE media_type SET visible = ? WHERE media_type = ? ";
 			}
 			if($entry_type == 'Relative Location'){
+				$select_exists_query = "SELECT loc_name FROM relt_location WHERE loc_name = ?";
 				$select_query = "SELECT sample_name FROM sample WHERE relt_loc_name = ?";
 				$update_query = "UPDATE relt_location SET visible = ? WHERE loc_name = ? ";
 			}
 			if($entry_type == 'rExtrKit'){
+				$select_exists_query = "SELECT r_kit_name FROM rna_extraction WHERE r_kit_name = ?";
 				$select_query = "SELECT sample_name FROM sample WHERE rna_extract_kit_name = ?";
 				$update_query = "UPDATE rna_extraction SET visible = ? WHERE r_kit_name = ? ";
 			}
 			
 			
 			if($entry_type == 'Sampler'){
+				$select_exists_query = "SELECT sampler_name FROM sampler WHERE sampler_name = ?";
 				$select_query = "SELECT sample_name FROM sample_sampler WHERE sampler_name = ?";
 				$update_query = "UPDATE sampler SET visible = ? WHERE sampler_name = ? ";
 			}
-			if($entry_type == 'Sensor'){//this has to be a different table
+			if($entry_type == 'Sensor'){	
+				$select_exists_query = "SELECT part_sens_name FROM particle_counter WHERE part_sens_name = ?";
 				$select_query = "SELECT daily_date FROM daily_data2_particle_counter WHERE part_sens_name = ?";
 				$update_query = "UPDATE particle_counter SET visible = ? WHERE part_sens_name = ? ";
 			}
+			
+			//check that the entry actually exists
+			$entry_exists_check = 'false';
+			if($stmt_exists = $dbc->prepare($select_exists_query)){
+				$stmt_exists -> bind_param('s', $entry_to_delete);
+	
+	  			if ($stmt_exists->execute()){
+	  			
+	    			$stmt_exists->bind_result($name);
+	    			while ($stmt_exists->fetch()){
+	        			echo "$name <br>";
+						$error = 'true';
+						$entry_exists_check = 'true';
+					}
+				} 
+				else {
+					$error = 'true';
+	    			die('Execute() failed: ' . htmlspecialchars($stmt_exists->error));
+				}
+			}
+			else {
+				$error = 'true';
+	    		die('Prepare() failed: ' . htmlspecialchars($stmt_exists->error));	
+			}
+			$stmt_exists -> close();
+			
 			
 				
 			//check samples exists for this project
@@ -106,14 +130,19 @@ $path = $_SERVER['DOCUMENT_ROOT'].$root; ?>
 				echo 'Do You Wish To Bulk Update The '.$entry_type.' For These Samples?';
 			}
 			
+			//Tell User Your Entry Does Not Exists
+			if($entry_exists_check == 'false'){
+				echo $entry_to_delete.' Does Not Exist. Please Check Name And Type';
+			}
+			
 			//update location visible to 0
 			$zero_visible = 0;
-		    if($error != 'true'){
+		    if($error == 'false'  && $entry_exists_check  == 'true'){
 		    	$stmt_delete = $dbc -> prepare($update_query);
 				$stmt_delete -> bind_param('is', $zero_visible,$entry_to_delete);
 				$stmt_delete -> execute();
 				$rows_affected_delete = $stmt_delete ->affected_rows;
-				if($rows_affected_delete == 0){ //return zero on update ...will return the same if did not update anything
+				if($rows_affected_delete == 0){ //return zero on update ...will return the same if did not update anything 
 					echo "You Successfully Deleted ".$entry_to_delete.'<br>';
 					$submitted = 'true';
 				}
